Loaders refactored and generilized
